pmb: only enable abuild-apk wrapper for buildroot (MR 2246)
Some packages like kernels are cross compiled from the native chroot.
the apk_wrapper won't work here! So only enable it for buildroot chroots

+++ pmb/build/init.py
@@ -61,7 +61,7 @@ def init(args, suffix="native"):
     apk_arch = pmb.parse.arch.from_chroot_suffix(args, suffix)
 
     # Add apk wrapper that runs native apk and lies about arch
-    if not os.path.exists(chroot + "/usr/local/bin/abuild-apk"):
+    if suffix.startswith("buildroot_") and not os.path.exists(chroot + "/usr/local/bin/abuild-apk"):
         with open(chroot + "/tmp/apk_wrapper.sh", "w") as handle:
             content = f"""
                 #!/bin/sh
